added lots of distributions

+++ examples/reference_rng.h
@@ -26,7 +26,7 @@ static inline uint64_t x256pp_next(uint64_t *s0, uint64_t *s1, uint64_t *s2,
 
 static inline double x256pp_u01(uint64_t *s0, uint64_t *s1, uint64_t *s2, uint64_t *s3) {
   uint64_t result = x256pp_next(s0, s1, s2, s3);
-  return (double)(result >> 11) * 0x1.0p-53;
+  return (result >> 11) * 0x1.0p-53;
 }
 
 static inline double x256pp_exponential(uint64_t *s0, uint64_t *s1, uint64_t *s2, uint64_t
@@ -47,7 +47,7 @@ static inline void fill_x256pp_u01(double *buf, int n, uint64_t *s0,
   uint64_t *s1, uint64_t *s2, uint64_t *s3) {
   for (int i = 0; i < n; i++) {
     uint64_t v = x256pp_next(s0, s1, s2, s3);
-    buf[i] = (double)(v >> 11) * 0x1.0p-53;
+    buf[i] = (v >> 11) * 0x1.0p-53;
   }
 }
 
@@ -68,8 +68,8 @@ static inline void fill_x256pp_norm_polar(double *buf, int n, uint64_t *s0,
       r1 = x256pp_next(s0, s1, s2, s3);
       r2 = x256pp_next(s0, s1, s2, s3);
       // u,v in (-1,1): 2*u01-1 from top 53 bits
-      u = 2.0*(double)(r1 >> 11) * 0x1.0p-53 - 1.0;
-      v = 2.0*(double)(r2 >> 11) * 0x1.0p-53 - 1.0;
+      u = 2.0*(r1 >> 11) * 0x1.0p-53 - 1.0;
+      v = 2.0*(r2 >> 11) * 0x1.0p-53 - 1.0;
       s = u*u + v*v;
     } while (s >= 1.0 || s == 0.0);
     double m = sqrt((-2.0*log(s))/s);
@@ -82,8 +82,8 @@ static inline void fill_x256pp_norm_polar(double *buf, int n, uint64_t *s0,
     do {
       r1 = x256pp_next(s0, s1, s2, s3);
       r2 = x256pp_next(s0, s1, s2, s3);
-      u = 2.0*(double)(r1 >> 11) * 0x1.0p-53 - 1.0;
-      v = 2.0*(double)(r2 >> 11) * 0x1.0p-53 - 1.0;
+      u = 2.0*(r1 >> 11) * 0x1.0p-53 - 1.0;
+      v = 2.0*(r2 >> 11) * 0x1.0p-53 - 1.0;
       s = u*u + v*v;
     } while (s >= 1.0 || s == 0.0);
     double m = sqrt((-2.0*log(s))/s);
